test(storage): split artifact store invariants
